+++ CHANGELOG.rst
@@ -1,4 +1,16 @@
-Release 3.3.0
+Release 4.1.0
+------------------------------------------------------------
+
+**Features**:
+
+**Bugs**:
+* Added namespace 1 to opc ua information model #18, by @huebl
+ 
+**Documentation**:
+
+
+
+Release 3.3.3
 ------------------------------------------------------------
 
 **Features**:
@@ -6,6 +18,32 @@ Release 3.3.0
 * refactoring application service interface #7, by @huebl
 
 **Bugs**:
+
+* Added advanced history access example #20, by @huebl
+ 
+**Documentation**:
+
+Release 3.3.2
+------------------------------------------------------------
+
+**Features**:
+
+**Bugs**:
+
+* Added namespace 1 to opc ua information model #18, by @huebl
+ 
+**Documentation**:
+
+
+
+Release 3.3.1
+------------------------------------------------------------
+
+**Features**:
+
+**Bugs**:
+
+* Replace ip address by hostname in application url #15, by @huebl
  
 **Documentation**: